+++ scrapper/src/test/java/edu/java/scrapper/database_test/jooq/JooqLinkRepositoryTest.java
@@ -1,4 +1,4 @@
-package edu.java.scrapper.database_test;
+package edu.java.scrapper.database_test.jooq;
 
 import edu.java.scrapper.IntegrationTest;
 import edu.java.scrapper.domain.repository.LinkRepository;
